feat(cli): add AReno ops coding agent
