+++ areno/agentic/coding/__init__.py
@@ -1,0 +1,22 @@
+"""Reusable coding-agent loop and tools."""
+
+from areno.agentic.coding.agent_loop import (
+    SYSTEM_PROMPT,
+    TOOLS,
+    initial_messages,
+    run_agentic_coding_loop,
+    run_conversation_turns,
+    run_single_task,
+)
+from areno.agentic.coding.coding_tools import CodingWorkspace, run_tool
+
+__all__ = [
+    "TOOLS",
+    "SYSTEM_PROMPT",
+    "CodingWorkspace",
+    "initial_messages",
+    "run_agentic_coding_loop",
+    "run_conversation_turns",
+    "run_single_task",
+    "run_tool",
+]
