+++ areno/agentic/coding/ops_knowledge.md
@@ -1,0 +1,140 @@
+# AReno Train/Serve Operations Knowledge
+
+You are running inside an existing AReno checkout and should use the local files
+and commands available in the current environment. Do not clone another AReno
+repo. Your job is to start a train or serve task successfully, inspect failures,
+and retry with adjusted parameters when the failure is likely recoverable.
+
+## Basic workflow
+
+1. Inspect the repository and examples before choosing commands:
+   - `pwd`
+   - `ls`
+   - `areno --help`
+   - `areno train --help`
+   - `areno serve --help`
+   - `areno check`
+   - `areno env`
+2. Inspect GPUs and memory before launching:
+   - `nvidia-smi`
+   - `nvidia-smi --query-gpu=index,name,memory.total,memory.used,memory.free --format=csv`
+   - `ps aux | grep -E "areno|python" | grep -v grep`
+   - `df -h . /tmp`
+3. Prefer a small smoke test first:
+   - one epoch or `--max-steps 1`
+   - small `--batch-size`
+   - small `--n-samples`
+   - conservative `--mini-bs`
+   - a valid `--save-path` only when checkpoint save must be tested
+4. Read the error message, adjust one or two parameters, and retry.
+5. Call `submit` only after the command is running or has completed successfully,
+   or when the task is blocked by missing files, missing GPUs, invalid API
+   credentials, or a non-recoverable dependency error.
+
+## Training command shape
+
+Common RL training command:
+
+```bash
+areno train \
+  --ckpt <model-or-local-checkpoint> \
+  --dataset-path <dataset> \
+  --dataset-loader-fn <loader.py> \
+  --reward-fn-path <reward.py> \
+  --algo gspo \
+  --world-size <gpu-count> \
+  --tp-size <tensor-parallel-size> \
+  --batch-size <prompts-per-step> \
+  --n-samples <samples-per-prompt> \
+  --mini-bs <train-microbatch> \
+  --max-running-prompts <rollout-concurrency> \
+  --max-steps 1
+```
+
+Useful examples:
+
+```bash
+areno train --ckpt Qwen/Qwen3.5-0.8B --dataset-path gsm8k:main \
+  --dataset-loader-fn examples/math/dataset_loader.py \
+  --reward-fn-path examples/math/math_verify_reward.py \
+  --algo gspo --world-size 1 --tp-size 1 --batch-size 1 --n-samples 2 \
+  --mini-bs 1 --max-running-prompts 2 --max-steps 1
+```
+
+```bash
+areno train --ckpt <local-ckpt> --dataset-path /home/admin/math/data \
+  --dataset-loader-fn examples/math/dataset_loader.py \
+  --reward-fn-path examples/math/math_verify_reward.py \
+  --algo gspo --world-size 8 --tp-size 4 --batch-size 32 --n-samples 8 \
+  --mini-bs 16 --max-running-prompts 256 --max-steps 1
+```
+
+Use `--save-path <dir> --save-interval 1 --max-steps 1` when the task asks to
+test checkpoint saving. Then test loading by using `--ckpt <dir>/step_000001`.
+
+## Serving command shape
+
+Common serve command:
+
+```bash
+areno serve --ckpt <model-or-local-checkpoint> --host 0.0.0.0 --port 8000 \
+  --world-size <gpu-count> --tp-size <tensor-parallel-size>
+```
+
+After serve starts, test it from another shell:
+
+```bash
+curl http://127.0.0.1:8000/v1/models
+curl http://127.0.0.1:8000/v1/chat/completions \
+  -H 'Content-Type: application/json' \
+  -d '{"model":"default","messages":[{"role":"user","content":"Say hi"}],"max_tokens":16}'
+```
+
+If the port is busy, choose another port and retry.
+
+## Memory tuning rules
+
+Rollout memory is mainly controlled by:
+
+- `--max-running-prompts`: higher means more concurrent rollout requests and
+  more KV cache memory.
+- `--max-new-tokens` and prompt length: longer sequences require more KV cache.
+- `--tp-size`: larger tensor parallel size usually lowers per-GPU model memory,
+  but changes the valid divisibility constraints for heads/layers.
+
+Training memory is mainly controlled by:
+
+- `--mini-bs`: higher means larger training microbatch and more activation
+  memory.
+- sequence length: longer rollout responses make train packs larger.
+- optimizer choice and whether rollout state is kept.
+
+If rollout OOM happens, reduce `--max-running-prompts`, `--batch-size`,
+`--n-samples`, or max sequence length. If train OOM happens, reduce `--mini-bs`
+first. If model loading OOM happens, increase `--tp-size` or use fewer other GPU
+processes.
+
+`--drop-rollout-state` means the rollout engine state is released before
+training to save memory. It can help when train OOM occurs after rollout. It may
+increase step overhead because rollout state must be rebuilt.
+
+## Recoverable failures and retries
+
+- CUDA out of memory during rollout: reduce `--max-running-prompts` by half.
+- CUDA out of memory during train: reduce `--mini-bs` by half.
+- OOM during startup/model loading: use a larger `--tp-size` if valid, or fewer
+  GPUs per process only if the model supports it.
+- `num_key_value_heads must be divisible by tp_size`: choose a `--tp-size` that
+  divides the model's key-value heads.
+- Port already in use for serve: retry with a different `--port`.
+- Dataset loader path missing: inspect `examples/` and choose the loader matching
+  the dataset.
+- Reward function missing for RL algorithms: provide `--reward-fn-path` or
+  `--reward-ckpt`.
+- SFT requires a dataset loader: provide `--dataset-loader-fn`.
+
+## Safety
+
+Do not run destructive cleanup commands except targeted cleanup under temporary
+directories when needed. Prefer inspecting disk usage before deleting anything.
+Do not kill unrelated user processes unless the task explicitly asks for it.
